Fixed checking process status on freebsd
